requirements.txt
annotated-types==0.7.0
asttokens==3.0.0
comm==0.2.2
debugpy==1.8.13
decorator==5.2.1
exceptiongroup==1.2.2
executing==2.2.0
ipykernel==6.29.5
ipython==8.35.0
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.7.2
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
packaging==24.2
parso==0.8.4
pexpect==4.9.0
platformdirs==4.3.7
ply==3.11
prompt_toolkit==3.0.50
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.11.3
pydantic_core==2.33.1
Pygments==2.19.1
python-dateutil==2.9.0.post0
PyYAML==6.0.2
pyzmq==26.4.0
simple-ddl-parser==1.7.1
six==1.17.0
stack-data==0.6.3
tornado==6.4.2
traitlets==5.14.3
typing-inspection==0.4.0
typing_extensions==4.13.1
wcwidth==0.2.13
ruamel.yaml